LayoutTests/editing/selection/ios/select-content-from-different-flow-005.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<meta charset="utf-8">
<head>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
    body {
        margin: 0;
    }

    #absolute {
        width: 100vw;
        height: 50px;
        position: absolute;
        left: 0;
        top: -100px;
    }

    #container {
        position: relative;
        margin-top: 150px;
    }
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("The highlight rects should be combined if the content is from same out of flow. Distinguish flows properly, positioned container + inline and block descendants.");

    var container = document.getElementById("container");
    await UIHelper.longPressElement(container);
    await UIHelper.waitForSelectionToAppear();
    getSelection().selectAllChildren(container);
    await UIHelper.waitForSelectionToAppear();
    selectionRects = await UIHelper.getUISelectionViewRects();

    shouldBe("selectionRects.length", "2");
    firstLineBottom = selectionRects[0].top + selectionRects[0].height;
    secondLineTop = selectionRects[1].top;
    shouldBe("firstLineBottom", "secondLineTop");

    finishJSTest();
});
</script>
</head>
<body>
    <div id="container">
        <div id="absolute">
            <span>Line 1</span>
            <div>Line 2<div>
        </div>
    </div>
</body>
</html>
